% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{env_action_space_info}
\alias{env_action_space_info}
\title{Get information (name and dimensions/bounds) of the environments's action space.}
\usage{
env_action_space_info(x, instance_id)
}
\arguments{
\item{x}{An instance of class "GymClient"; this object has "remote_base" as an attribute.}

\item{instance_id}{A short identifier (such as "3c657dbc") for the environment instance.}
}
\value{
A list containing "name" (such as "Discrete"), and additional dimensional info (such as "n") which varies from space to space.
}
\description{
Get information (name and dimensions/bounds) of the environments's action space.
}

