% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{env_close}
\alias{env_close}
\title{Flush all monitor data to disk.}
\usage{
env_close(x, instance_id)
}
\arguments{
\item{x}{An instance of class "GymClient"; this object has "remote_base" as an attribute.}

\item{instance_id}{A short identifier (such as "3c657dbc") for the environment instance.}
}
\value{
NULL.
}
\description{
Flush all monitor data to disk.
}
\examples{
\dontrun{
remote_base <- "http://127.0.0.1:5000"
client <- create_GymClient(remote_base)
env_id <- "CartPole-v0"
instance_id <- env_create(client, env_id)
env_close(client, instance_id)
}
}

