% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{GymClient}
\alias{GymClient}
\title{Create a GymClient instance.}
\usage{
GymClient(remote_base)
}
\arguments{
\item{remote_base}{The URL of the OpenAI  gym server. This value is usually "http://127.0.0.1:5000".}
}
\value{
An instance of class "GymClient"; this object has "remote_base" as an attribute.
}
\description{
This function instantiates a GymClient instance to integrate with an OpenAI Gym server.
}
\examples{
\dontrun{
remote_base <- "http://127.0.0.1:5000"
client <- GymClient(remote_base)
}
}

