% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{print.GymClient}
\alias{print.GymClient}
\title{Represent a GymClient instance on the command line.}
\usage{
\method{print}{GymClient}(x, ...)
}
\arguments{
\item{x}{An instance of class "GymClient"; this object has "remote_base" as an attribute.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
x A GymClient instance.
}
\description{
Represent a GymClient instance on the command line.
}

