% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{random_discrete_agent}
\alias{random_discrete_agent}
\title{A sample random discrete agent.}
\usage{
random_discrete_agent(n)
}
\arguments{
\item{n}{The number of discrete action spaces available.}
}
\value{
NULL.
}
\description{
A sample random discrete agent.
}

