% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{env_close}
\alias{env_close}
\title{Flush all monitor data to disk.}
\usage{
env_close(x, instance_id)
}
\arguments{
\item{x}{An instance of class "GymClient"; this object has "remote_base" as an attribute.}

\item{instance_id}{A short identifier (such as "3c657dbc") for the environment instance.}
}
\value{
NULL.
}
\description{
Flush all monitor data to disk.
}

