% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_server_error_or_raise_for_status}
\alias{parse_server_error_or_raise_for_status}
\title{TO BE EDITED.}
\usage{
parse_server_error_or_raise_for_status(response)
}
\value{
TO BE EDITED.
}
\description{
TO BE EDITED.
}

